Remove dependency of executable_worker -> test_utils
